refactor: moved review component to Show dir
